+++ com.unity.mobile.notifications/Editor/NotificationSettings.cs
@@ -182,6 +182,7 @@ namespace Unity.Notifications
             internal static readonly string PUSH_NOTIFICATION_PRESENTATION = "UnityRemoteNotificationForegroundPresentationOptions";
             internal static readonly string USE_APS_RELEASE = "UnityUseAPSReleaseEnvironment";
             internal static readonly string USE_LOCATION_TRIGGER = "UnityUseLocationNotificationTrigger";
+            internal static readonly string ADD_TIME_SENSITIVE_ENTITLEMENT = "UnityAddTimeSensitiveEntitlement";
 
             /// <summary>
             /// It's recommended to make the authorization request during the app's launch cycle. If this is enabled the user will be shown the authorization pop-up immediately when the app launches. If it’s unchecked you’ll need to manually create an AuthorizationRequest before your app can send or receive notifications.
@@ -285,6 +286,21 @@ namespace Unity.Notifications
                 set
                 {
                     SetSettingValue<bool>(BuildTargetGroup.iOS, USE_LOCATION_TRIGGER, value);
+                }
+            }
+
+            /// <summary>
+            /// Add entitlement to enable notifications with time-sensitive interruption level.
+            /// </summary>
+            public static bool AddTimeSensitiveEntitlement
+            {
+                get
+                {
+                    return GetSettingValue<bool>(BuildTargetGroup.iOS, ADD_TIME_SENSITIVE_ENTITLEMENT);
+                }
+                set
+                {
+                    SetSettingValue<bool>(BuildTargetGroup.iOS, ADD_TIME_SENSITIVE_ENTITLEMENT, value);
                 }
             }
         }